requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==1.6.0
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg==3.1.10
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
